Pull common JS to common.js.
